NhacCuaTuiScraper/requirements.txt
scrapy
scrapy-splash
mysql-connector-python
boto3
